chore: replace gsheets icon and rename file name
